% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network.R
\name{get_network}
\alias{get_network}
\title{flowchart of R projects}
\usage{
get_network(dir, variations = c(" <- function", "<- function", "<-function"),
  pattern = "\\\\.R")
}
\arguments{
\item{dir}{a path that includes the functions}

\item{variations}{a character vector with the function's definition string.
The default is c(" <- function", "<- function", "<-function").}

\item{pattern}{a string with the file suffix - default is "\\.R".}
}
\value{
Returns an object with the adjacency matrix \code{$matrix} and 
  and igraph object \code{$igraph}.
}
\description{
With this function a network plot of the connections of the functions
in a given path can be created.
}
\note{
TODO: list with exclude files and comments ' ' in one line
}
\examples{
\dontrun{
net <- get_network(dir = "R/")
g1 <- net$igraph
plot(g1)
}

}
\seealso{
For more information see 
\href{https://www.statworx.com/de/blog/flowcharts-of-functions/}{our blog}.
}
\author{
Jakob Gepp
}
