% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statworx.R
\docType{package}
\name{statworx-package}
\alias{statworx-package}
\alias{statworx}
\alias{helperfunctions}
\title{This package contains some helper functions.}
\description{
There is no other purpose to the whole package but to collect and distribute
some of the most used functions or helper functions which are used over and
over again.
}
\author{
Jakob Gepp <jakob.gepp@statworx.com>
}
